pyproject.toml
[project]
name = "nt-gov-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.17.0",
    "django>=5.2.5",
    "faiss-cpu>=1.12.0",
    "grandalf>=0.8",
    "ipython>=9.5.0",
    "json-repair>=0.50.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
]
